refactor: move default presets from server to engine
