+++ packages/server/src/plugin-manager/index.ts
@@ -1,3 +1,2 @@
 export * from './clientStaticUtils';
 export * from './plugin-manager';
-export * from './plugin-presets';
